+++ final-project-be/Dtos/Post/PostDto.cs
@@ -11,6 +11,7 @@ namespace final_project_be.Dtos.Post
         public int SubCategoryId { get; set; }
         public string Title { get; set; }
         public string Content { get; set; }
+        public int ReportCount { get; set; }
         public List<PostFileDto>? PostFiles { get; set; }
         public List<CommentDto>? Comments { get; set; }
     }
