+++ final-project-be/Dtos/User/UserProfileDto.cs
@@ -10,5 +10,6 @@
         public DateTime? Birthday { get; set; }
         public string Gender { get; set; }
         public string Address { get; set; }
+        public string Avatar { get; set; }
     }
 }
