+++ final-project-be/Service/Mapping/MapperProfile.cs
@@ -36,7 +36,6 @@ namespace final_project_be.Service.Mapping
             CreateMap<UserProfileDto, User>().ReverseMap();
             CreateMap<UserProfileDto, UserMetadata>().ReverseMap();
             CreateMap<PostFileDto, PostFile>().ReverseMap();
-            CreateMap<PostManagerDto, Post>().ReverseMap();
             CreateMap<PostCreateDto, Post>().ReverseMap();
             CreateMap<PostFileCreateDto, Post>().ReverseMap();
         }
