+++ final-project-be/appsettings.Development.json
@@ -13,6 +13,6 @@
     "ExpirationDays": 3
   },
   "ConnectionStrings": {
-    "MyConnection": "Server=tcp:final-project-bedbserver.database.windows.net,1433;Initial Catalog=final-project-be_db;Persist Security Info=False;User ID=hoangkaito03;Password=Nguyenthanhhoang03;MultipleActiveResultSets=False;Encrypt=True;TrustServerCertificate=False;Connection Timeout=30;"
+    "MyConnection": "server=localhost; database =Final-project-BE;uid=admin;pwd=123;Trust Server Certificate=true"
   }
 }
